backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
onnxruntime==1.23.0
opencv-python==4.10.0.84
numpy==2.3.3
pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.0
google-generativeai==0.7.1
deepface==0.0.91
